pyproject.toml
[project]
  name = "job-search-agent"
  version = "0.1.0"
  requires-python = ">=3.11"
  dependencies = [
      "anthropic>=0.40.0",
      "pydantic>=2.0.0",
      "asyncpg>=0.29.0",
      "hydra-core>=1.3.0",
      "omegaconf>=2.3.0",
      "pymupdf>=1.24.0",
      "apify-client>=1.6.0",
      "httpx>=0.27.0",
      "python-dotenv>=1.0.0",
      "pytest>=9.0.3",
      "pytest-asyncio>=1.3.0",
      "pytest-cov>=7.1.0",
]

  [project.optional-dependencies]
  dev = [
      "pytest>=8.0.0",
      "pytest-asyncio>=0.23.0",
      "pytest-cov>=5.0.0",
  ]

  [tool.pytest.ini_options]
  asyncio_mode = "auto"
  testpaths = ["tests"]
